requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
langchain-text-splitters
langchain-huggingface
sentence_transformers
tensorflow
huggingface-hub
pandas
faiss-cpu
langchain-groq
tf-keras
langchain-chroma
streamlit
numexpr
youtube-transcript-api
pytube
yt-dlp
validators